RetroPlayer: Move buffers to new buffers/ folder
