+++ xbmc/cores/RetroPlayer/buffers/BaseRenderBuffer.cpp
