+++ xbmc/cores/RetroPlayer/buffers/BaseRenderBuffer.h
