+++ xbmc/cores/RetroPlayer/buffers/BaseRenderBufferPool.cpp
