+++ xbmc/cores/RetroPlayer/buffers/BaseRenderBufferPool.h
