+++ xbmc/cores/RetroPlayer/buffers/IRenderBuffer.h
