+++ xbmc/cores/RetroPlayer/buffers/IRenderBufferPool.h
