+++ xbmc/cores/RetroPlayer/buffers/RenderBufferManager.h
