+++ xbmc/cores/RetroPlayer/buffers/video/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(SOURCES RenderBufferGuiTexture.cpp
+            RenderBufferSysMem.cpp
+)
+
+set(HEADERS RenderBufferGuiTexture.h
+            RenderBufferSysMem.h
+)
+
+core_add_library(rp-buffers-video)
