+++ xbmc/cores/RetroPlayer/buffers/video/RenderBufferGuiTexture.cpp
