+++ xbmc/cores/RetroPlayer/buffers/video/RenderBufferSysMem.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "BaseRenderBuffer.h"
+#include "cores/RetroPlayer/buffers/BaseRenderBuffer.h"
 
 #include <stdint.h>
 #include <vector>
