+++ xbmc/cores/RetroPlayer/process/CMakeLists.txt
@@ -1,19 +1,7 @@
-set(SOURCES BaseRenderBuffer.cpp
-            BaseRenderBufferPool.cpp
-            RenderBufferGuiTexture.cpp
-            RenderBufferManager.cpp
-            RenderBufferSysMem.cpp
-            RPProcessInfo.cpp
+set(SOURCES RPProcessInfo.cpp
 )
 
-set(HEADERS BaseRenderBuffer.h
-            BaseRenderBufferPool.h
-            IRenderBuffer.h
-            IRenderBufferPool.h
-            RenderBufferGuiTexture.h
-            RenderBufferManager.h
-            RenderBufferSysMem.h
-            RPProcessInfo.h
+set(HEADERS RPProcessInfo.h
 )
 
 core_add_library(rp-process)
