+++ xbmc/cores/RetroPlayer/process/rbpi/RenderBufferMMAL.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "cores/RetroPlayer/process/BaseRenderBuffer.h"
+#include "cores/RetroPlayer/buffers/BaseRenderBuffer.h"
 
 #include <interface/mmal/mmal.h>
