+++ xbmc/cores/RetroPlayer/rendering/VideoRenderers/RPRendererGuiTexture.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "RPBaseRenderer.h"
-#include "cores/RetroPlayer/process/BaseRenderBufferPool.h"
+#include "cores/RetroPlayer/buffers/BaseRenderBufferPool.h"
 #include "cores/RetroPlayer/process/RPProcessInfo.h"
 #include "cores/IPlayer.h"
